+++ development/src/GXDLMSAssociationLogicalName.cpp
@@ -571,7 +571,7 @@ int CGXDLMSAssociationLogicalName::Invoke(CGXDLMSSettings& settings, CGXDLMSValu
     if (e.GetIndex() == 1)
     {
         int ret;
-        unsigned long ic = 0;
+        uint32_t ic = 0;
         CGXByteBuffer secret;
         CGXByteBuffer* readSecret;
         if (settings.GetAuthentication() == DLMS_AUTHENTICATION_HIGH_GMAC)
